Build functionality / Discord.js framework (rewrite) (#44)
# Build functionality / Discord.js Framework (rewrite)

In this iteration we introduced cache invalidation with a simple yet effective approach. Although we soon plan on dropping support for Discord.js and Eris (soon to be added - another PR) in favor for our own library which interacts with the Discord API, we have rewritten the Discord.js mini framework we built into our codebase. Our own Discord API library will allow us to have more control over how the user experience is. This is also an improvement of the last iteration, we added caching (previously mentioned), more validation (catches more edge cases), and is more reliant on our dist directory ecosystem (manifests, and \`cache\` directory 🎉) instead of the file contents and its "use" case statement (poor decision).

> To those who married Discord.js and Eris: When we drop Discord.js as the default framework, we will allow you to opt-in of Discord.js in `jujutsu.config.js`. And for the Eris couples, in that future iteration, we will create a Eris framework for those who want to use Eris.

⚠️ No predicted date of this feature

This is most likely the biggest change that will be implemented before the initial release. Also, for those wondering how this is a framework for building Discord bots, while most of the code has nothing to do with the Discord API --> This is a tool that can allow bots, both big and small to easily scale their Discord bot. This also relieves the developer of deciding on trivial things (like where to place commands, events, etc.). This "framework" also allows the developer to focus more on developing their Discord bot where is matter: interactions and commands, than the background code (ex: command & event managers). We derive the framework from the part that it extends Discord.js, Eris (and soon to be our codebase).

## Keynote
If you didn't read through my lengthy rant (🤣), here are two senetences explaining what I said:

Jujutsu.js will allow any Discord bot developer, learning or experienced, to quickly develop a Discord bot with the hassle or worry of the "backend of the backend" of the Discord API. This will, in short, allow them to focus on the user experience, rather than maintaining the "backend of the backend".

Join the Discord: [soon™️]()

### TODO

- [X] ~Validate event information~
- [x] Add subcommands to app build manifest
- [X] ~Validate command information~
- [X] ~`jujutsu build` not working as expected~
- [X] ~Add caching for incremental building~
- [X] ~Rewrite the Discord.js Managers to be up to date with new build functionality~
- [X] ~Rewrite how the server identifies commands and events~
- [x] Add more validation other than file exports
- [X] Implement `app` directory event.js and command.js into runtime (still built, just not used)
- [X] Delete old transformed file in `.jujutsu/server` when a file is updated 
(unique ids stop overwrites --> old files leftover)
- [x] ~Fix bug when name in build manifest changes when inc. build runs~
- [x] ~Delete cache and 'chunk' of deleted event files (both in `events` and `app` directory)~
- [x] ~Check if command.js file with a depth of 2 has a parent command.js, if not --> ignore subcommand~
- [X] ~Add more slash command options~ (need to add more still)
- [x] ~Add subcommands to Discord.js framework~

Co-authored-by: kijv <93486795+kijv@users.noreply.github.com>